Move filter, sort and fetch logic to VmsList component
