+++ src/components/VmsList/style.css
@@ -51,8 +51,6 @@ td .vm-status {
     margin-bottom: 0;
 }
 
-
-
 .vm-detail-link {
     color: black;
 }
